token class implemented
added the token class. double check me to make sure we have all bnf versions of the token, in addition to this i have a basic tokenizer set up which grabs only the first token. need to define states and methods for grabbing tokens, but the backbone of setting the tokens should be done.